requirements.txt
pandas
python-pptx
openai
python-dotenv
langchain-core
langchain
langchain-openai
langchain-community